update to KinglittleQ/SuperPoint_SLAM

+++ src/MapPoint.cc
@@ -19,7 +19,7 @@
 */
 
 #include "MapPoint.h"
-#include "ORBmatcher.h"
+#include "SPmatcher.h"
 
 #include<mutex>
 
@@ -278,20 +278,20 @@ void MapPoint::ComputeDistinctiveDescriptors()
         Distances[i][i]=0;
         for(size_t j=i+1;j<N;j++)
         {
-            int distij = ORBmatcher::DescriptorDistance(vDescriptors[i],vDescriptors[j]);
+            float distij = ORBmatcher::DescriptorDistance(vDescriptors[i],vDescriptors[j]);
             Distances[i][j]=distij;
             Distances[j][i]=distij;
         }
     }
 
     // Take the descriptor with least median distance to the rest
-    int BestMedian = INT_MAX;
+    float BestMedian = INT_MAX;
     int BestIdx = 0;
     for(size_t i=0;i<N;i++)
     {
-        vector<int> vDists(Distances[i],Distances[i]+N);
+        vector<float> vDists(Distances[i],Distances[i]+N);
         sort(vDists.begin(),vDists.end());
-        int median = vDists[0.5*(N-1)];
+        float median = vDists[0.5*(N-1)];
 
         if(median<BestMedian)
         {
